Centralize EstimatePerVertexCost and Execute_BlockTransferStart
