+++ GPU/Vulkan/GPU_Vulkan.h
@@ -127,7 +127,6 @@ public:
 	void Execute_TgenMtxData(u32 op, u32 diff);
 	void Execute_BoneMtxNum(u32 op, u32 diff);
 	void Execute_BoneMtxData(u32 op, u32 diff);
-	void Execute_BlockTransferStart(u32 op, u32 diff);
 
 	// Using string because it's generic - makes no assumptions on the size of the shader IDs of this backend.
 	std::vector<std::string> DebugGetShaderIDs(DebugShaderType shader) override;
